Implemented generic selector
