+++ custom_components/hass_pontos/device_config/conf_pontos.py
@@ -239,3 +239,10 @@ BUTTONS = {
         "availability_sensor": "alarm_status"
     }
 }
+
+SELECT_CONFIG = {
+    "profile_select": {
+        "name": "Profile select",
+        "type": "profile_select"
+    }
+}
