+++ custom_components/hass_pontos/device_config/conf_safetech_v4.py
@@ -244,3 +244,10 @@ BUTTONS = {
         "availability_sensor": "alarm_status"
     }
 }
+
+SELECT_CONFIG = {
+    "profile_select": {
+        "name": "Profile select",
+        "type": "profile_select"
+    }
+}
