+++ custom_components/hass_pontos/device_config/conf_trio.py
@@ -230,3 +230,11 @@ BUTTONS = {
         "availability_sensor": "alarm_status"
     }
 }
+
+SELECT_CONFIG = {
+    "profile_select": {
+        "name": "Profile select",
+        "type": "profile_select"
+    }
+}
+
