+++ custom_components/hass_pontos/translations/en.json
@@ -202,7 +202,7 @@
                   "mode_standard": "Standard",
                   "mde_eco": "ECO",
                   "mode_power": "Power",
-                  "mode_automatic": "automatic"
+                  "mode_automatic": "Automatic"
               }
           },
           "regeneration_status": {
